extracts logic for converting fields to properties
New class is called FieldConverter and contains field2property and
it's dependencies
